pto refactoring
